Added new method "getPageRevisions" to Mediawiki.
The "getPages" methods return only the latest revision of each page (probably because the "rvlimit" "may only be used with a single page").

The has basically the same logic as the "getPageContent"-method, just that it returns a list of revision objects rather than the content string. It is minimally parameterized with the "rvlimit" and the "rvprop", leaving a backdoor for anything else via arbitrary "queryParams".

See https://www.mediawiki.org/wiki/API:Revisions for what query params would actually work.

Added a rudimentary test that checks the consistency with the results of existing methods.

Wasn't sure about the Java version, so I sticked to 7, just in case.